requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
Markdown==3.2.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
